Move ToastContext and utilities to shared components
