+++ packages/shared-components/src/core/utils/ToastContext.tsx
@@ -1,9 +1,9 @@
 /*
-Copyright 2024 New Vector Ltd.
-Copyright 2024 The Matrix.org Foundation C.I.C.
-
-SPDX-License-Identifier: AGPL-3.0-only OR GPL-3.0-only OR LicenseRef-Element-Commercial
-Please see LICENSE files in the repository root for full details.
+* Copyright 2024 New Vector Ltd.
+* Copyright 2024 The Matrix.org Foundation C.I.C.
+*
+* SPDX-License-Identifier: AGPL-3.0-only OR GPL-3.0-only OR LicenseRef-Element-Commercial
+* Please see LICENSE files in the repository root for full details.
 */
 
 import { type ReactNode, createContext, useCallback, useContext, useEffect, useState, useMemo } from "react";
